requirements.txt
itsdangerous
amqp==2.5.0
attrs==19.1.0
billiard==3.6.0.0
celery==4.3.0
celery-progress==0.0.4
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
demoji==0.1.5
Django==2.2.4
djangorestframework==3.10.2
dnspython==1.16.0
drf-yasg==1.16.1
eventlet==0.25.1
greenlet==0.4.15
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
kombu==4.6.3
MarkupSafe==1.1.1
monotonic==1.5
natural==0.2.0
packaging==19.1
psycopg2==2.8.3
pyparsing==2.4.2
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.2
raven==6.10.0
redis==3.3.8
requests==2.22.0
ruamel.yaml==0.16.1
ruamel.yaml.clib==0.1.2
sentry-sdk==0.10.2
shortuuid==0.5.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
